Disabling USART during sleep power down
My usage of RF69::pcIntCount may be suspect

+++ RF69.h
@@ -19,7 +19,7 @@ namespace RF69 {
     extern uint16_t underrun;
     extern uint8_t  present;
     extern uint16_t pcIntCount;
-    extern uint8_t  pcIntBits;
+    extern uint16_t  pcIntBits;
 
     void setFrequency (uint32_t freq);
     bool canSend ();
